+++ packages/spacecat-shared-athena-client/src/traffic-analysis/traffic-data-base-response.js
@@ -34,6 +34,7 @@ export const TrafficDataResponseDto = {
    */
   toJSON: (data) => ({
     week: data.week,
+    month: data.month,
     type: data.trf_type,
     channel: data.trf_channel,
     platform: data.trf_platform,